FIX BUG: stability plot giving different results for same element.
The stability plot for element X was different if done fresh or after plotting the stability of another element.

Explanation: Since python lists and dicts are pointers for a memory space, and doing a = b means both point at the same memory space, when varying one element to calculate stability over a range, Simcav was actually modifying the original "element_list". Doing list_b = list(list_a) did NOT solve it because the dicts inside the lists where still pointing to the same memory space (ie. a is b -> False, but a[0] is b[0] -> True).

Solution: The solution was to rewrite the list element (ie. b[i]=newdict). This removes certain keys from the element's dictionary, but they aren't needed in the following calculations.

This was also a problem in the beam_size calculation. Solved in the same way.